+++ templates/evidence.fields
@@ -4,4 +4,10 @@ issue.location
 issue.severity
 issue.confidence
 issue.request
+issue.request_1
+issue.request_2
+issue.request_3
 issue.response
+issue.response_1
+issue.response_2
+issue.response_3
